# backend/api-first/openapi.yaml
openapi: 3.1.0
info:
  title: KOSMO API
  version: "v1"
  description: >-
    API de KOSMO - plataforma low-code de spec-driven development (SDD) asistida por IA.
    Ciclo completo: Descubrimiento -> Características -> Requisitos -> Modelo -> Código.

    Convenciones:
    - Autenticación PKCE (RFC 7636) con JWT RS256 (bearerAuth).
    - Errores RFC 7807 (application/problem+json) con type URN, trace_id y violations, o OAuthErrorResponse (RFC 6749) / HttpErrorResponse.
    - Fechas ISO-8601 UTC con sufijo Z.
    - Las fases siguen el orden Descubrimiento -> Características -> Requisitos -> Modelo -> Código.
      La consistencia se evalúa exclusivamente hacia la derecha (downstream).
  contact:
    name: Equipo KOSMO
    email: dev@kosmo.app
    url: https://github.com/CesarPantoja1/KOSMO
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local — desarrollo en máquina del programador
  - url: https://api-dev.kosmo.app
    description: Desarrollo — entorno de integración continua
  - url: https://api.kosmo.app
    description: Producción — tráfico real de usuarios
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/features/{feature_id}/requirements/generate:
    post:
      tags: [requirements]
      summary: Generar requisitos EARS con IA
      description: Genera requisitos EARS para la caracteristica indicada.
      operationId: requirements_generate
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdRateLimitQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequirementsRequest"
      responses:
        "200":
          description: Requisitos generados exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateRequirementsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/features/{feature_id}/requirements:
    get:
      tags: [requirements]
      summary: Obtener requisitos de una caracteristica
      description: Recupera los requisitos en formato Markdown asociados a la caracteristica.
      operationId: requirements_get
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdQuery"
      responses:
        "200":
          description: Documento de requisitos en Markdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequirementsDocumentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [requirements]
      summary: Guardar/actualizar requisitos
      description: Actualiza el documento Markdown de requisitos de la caracteristica especificada.
      operationId: requirements_save
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRequirementsRequest"
      responses:
        "200":
          description: Requisitos guardados exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  feature_id:
                    type: string
                  message:
                    type: string
                    example: ok
                required: [feature_id, message]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [requirements]
      summary: Eliminar requisitos
      description: Elimina el documento de requisitos EARS de la caracteristica especificada.
      operationId: requirements_delete
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdQuery"
      responses:
        "200":
          description: Requisitos eliminados exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  feature_id:
                    type: string
                required: [status, feature_id]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/features/{feature_id}/requirements/refine:
    post:
      tags: [requirements]
      summary: Refinar requisitos EARS con IA
      description: >-
        Refina los requisitos de una caracteristica aplicando las instrucciones del usuario.
        Requiere que la caracteristica ya tenga requisitos generados.
      operationId: requirements_refine
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdRateLimitQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefineRequirementsRequest"
      responses:
        "200":
          description: Requisitos refinados exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefineRequirementsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/features/{feature_id}/requirements/regenerate:
    post:
      tags: [requirements]
      summary: Regenerar requisitos EARS con IA
      description: >-
        Regenera los requisitos de una caracteristica a partir del estado actual de la caracteristica padre,
        manteniendo la estructura EARS y criterios de aceptacion Dado-Cuando-Entonces.
      operationId: requirements_regenerate
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdQuery"
      responses:
        "200":
          description: Requisitos regenerados exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateRequirementsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"

  # ══════════════════════════════ REQUIREMENT CHAT ══════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 emitido por /api/v1/auth/token o /api/v1/auth/refresh
  parameters:
...
    FeatureId:
      name: feature_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador o slug de la característica
    ProjectIdQuery:
      name: project_id
      in: query
      required: true
      schema:
        type: string
      description: Identificador del proyecto
...
    ProjectIdRateLimitQuery:
      name: project_id
      in: query
      required: false
      schema:
        type: string
        default: ""
      description: Identificador opcional del proyecto para aplicar rate limiting cuando se invoca fuera de una ruta anidada de proyecto.
  responses:
    Unauthorized:
      description: Token de acceso ausente, invalido, expirado o revocado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HttpErrorResponse"
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    BadGateway:
      description: Error al invocar el servicio de IA o LLM
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
    ProblemDetail:
      type: object
      description: Error RFC 7807 (application/problem+json)
      properties:
        type:
          type: string
          description: URN del error (ej. urn:kosmo:projects:not-found)
          example: urn:kosmo:projects:not-found
        title:
          type: string
          description: Titulo corto del error
        status:
          type: integer
          description: Codigo de estado HTTP
        detail:
          type: string
          description: Descripcion detallada del error
        instance:
          type: string
          description: Ruta o recurso que origino el error
        trace_id:
          type: string
          description: Identificador de correlacion distribuida (X-Trace-Id)
        violations:
          type: array
          items:
            type: object
            properties:
              loc:
                type: string
                description: Ruta al campo invalido
              msg:
                type: string
                description: Mensaje de error de validacion
              input:
                description: Valor recibido que causo la violacion
      required: [type, title, status, detail]
    HttpErrorResponse:
      type: object
      description: Respuesta de error generica para errores HTTP (4xx/5xx)
      properties:
        detail:
          type: string
          description: Mensaje descriptivo del error
          example: Recurso no encontrado.
      required: [detail]

    # ── Auth ──
...
    GenerateRequirementsRequest:
      type: object
      additionalProperties: false
      properties:
        project_id:
          type: string
          description: ID del proyecto
      required: [project_id]
    EARSRequirementView:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        statement:
          type: string
        pattern:
          type: string
          description: Patron EARS (ubiquitous, event_driven, state_driven, optional, unwanted, complex)
        acceptance_criteria:
          type: array
          items:
            type: object
      required: [id, title, statement, pattern]
    GenerateRequirementsResponse:
      type: object
      properties:
        project_id:
          type: string
        feature_id:
          type: string
        feature_number:
          type: integer
        document_markdown:
          type: string
        total:
          type: integer
        requirements:
          type: array
          items:
            $ref: "#/components/schemas/EARSRequirementView"
      required: [project_id, feature_id, feature_number, document_markdown, total, requirements]
    RequirementsDocumentResponse:
      type: object
      properties:
        document_markdown:
          type: string
        total:
          type: integer
      required: [document_markdown, total]
    SaveRequirementsRequest:
      type: object
      additionalProperties: false
      properties:
        project_id:
          type: string
        markdown:
          type: string
      required: [project_id, markdown]
    RefineRequirementsRequest:
      type: object
      additionalProperties: false
      properties:
        project_id:
          type: string
        instructions:
          type: string
          minLength: 1
          maxLength: 500
      required: [project_id, instructions]
    RefineRequirementsResponse:
      type: object
      properties:
        feature_id:
          type: string
        feature_number:
          type: integer
        document_markdown:
          type: string
        total:
          type: integer
      required: [feature_id, feature_number, document_markdown, total]
    RegenerateRequirementsResponse:
      type: object
      properties:
        artifact_id:
          type: string
        content:
          type: string
        phase:
          type: string
      required: [artifact_id, content, phase]

    # ── Modelo ──
...